Early Implementation of GUI
Opens a window that allows for strings to be entered, which will be
mapped to keybinds from a dropdown menu.
Also, a button can be pressed to view all keybinds.

Note that due to github codespaces not having environments.
testgui.py must be run on your host computer.

gui.py is a failed attempt at a gui using pysimplegui